requirements.txt
# Django
Django==4.2.7
djangorestframework==3.14.0

# Authentication
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9

# Environment variables
python-decouple==3.8

# Image handling (AWS S3 a venir)
Pillow==10.1.0

# API Documentation
drf-spectacular==0.26.5

django-filter==23.5
django-cors-headers==4.3.1
